requirements.txt
# Requires Python >= 3.12
langchain
langchain-groq
python-dotenv
langchain-tavily
langchain-community
langgraph
langchain
wikipedia
fastapi
uvicorn
pydantic
langsmith
